convert `strategy.py` into a package and split into files (should be backwards-compatible)
